Move wiki samples out of the main junit dir
