+++ javaparser-testing/src/test/java/com/github/javaparser/junit/wiki_samples/MethodPrinter.java
@@ -1,4 +1,4 @@
-package com.github.javaparser.junit;
+package com.github.javaparser.junit.wiki_samples;
 
 import com.github.javaparser.JavaParser;
 import com.github.javaparser.ast.CompilationUnit;
