+++ javaparser-testing/src/test/java/com/github/javaparser/junit/wiki_samples/PrintingTheCompilationUnitToSystemOutputTest.java
@@ -1,12 +1,13 @@
-package com.github.javaparser.junit;
+package com.github.javaparser.junit.wiki_samples;
 
 import org.junit.Test;
 
-public class ChangingMethodsFromAClassWithAVisitorTest {
+public class PrintingTheCompilationUnitToSystemOutputTest {
     @Test
     public void printingTheCompilationUnitToSystemOutput() throws Exception {
         try (TestFileToken f = new TestFileToken("test.java")) {
-            MethodChanger_1.main(new String[]{});
+            CuPrinter.main(new String[]{});
         }
     }
+
 }
