+++ javaparser-testing/src/test/java/com/github/javaparser/junit/wiki_samples/VisitingClassMethodsTest.java
@@ -1,4 +1,4 @@
-package com.github.javaparser.junit;
+package com.github.javaparser.junit.wiki_samples;
 
 import org.junit.Test;
